Now got the java app to output an array of condensed song data.
This is then copied into songData.h for the avr code (but this
could be automated later e.g. file creation and stuff).
The format of the data is a little confusing just to save memory:

It is a 4 byte, long-style variable
First byte is the period of the note in "ticks"
Next 22 bits are for the tick count (i.e. when the note should be played)
Final 2 bits are for the floppy/track number (0 -> 3)

Next to add the decoding and song control parts to avr code