Got a working sketch to play a portion of a midi file.
The length of the song data array is a real issue. It currently resides in SRAM which is very restriced. A possible
fix for this is to move some of it into eeprom and swap them out in pages(?). that would only give space for 256 more
notes (+50%)...
The best outcome would be to somehow store it in the flash memory but I'm not sure if that is achieveable.

Looping through each floppy device have been fixed. the sizeof(array) doesn't return the number of elements in the
array, but the number of bytes it takes up in memory. So sizeof(_drives)/sizeof(Floppy) can be used to determine the
number of elements in the _drives array.

Next steps are to more efficiently store the song data and begin implementing the alarm clock functionality