Added "pgmspace.h" library for storing and retriving the song data from the flash memory (program space).
As the actual program is currently around 7 kB, there is then 23 kB left for song data. Each note is packed
into an unsigned long (4 bytes) so that is space for almost 6000 notes! (mm2 metalman only has 1400)

Because of this change, the length of the songData array can no longer be easily determined so a definition of
its length has been added in the songData.h header. The java app will need to be amended accordingly to output
this automatically, as well as the include for the pgmspace library.