Moved the floppy drive pins to work on  port register B (pins 8->13) to accomodate for the external interrupt pins being 2 and 3 which are needed for user control.
Adapted the java midi parser to output the includes, definitions and such required for it to
compile with no editing. (main java file is called "main" not "arrayparser" but i wont
remove the undeeded file so git can be used to revert to an earlier version.

Next step is to add some user control for starting/stopping the song and then finally adding
in the alarm clock functionality.